# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "This is the OpenAPI specification for your project."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "SeasonResponse": {
        "type": "object",
        "properties": {
          "id": {
            "allOf": [
              {
                "$ref": "#/components/schemas/idInt"
              }
            ],
            "description": "Season numeric identifier (PK)"
          },
          "created_at": {
            "allOf": [
              {
                "$ref": "#/components/schemas/timestamp"
              }
            ],
            "description": "Row creation time in UTC (ISO 8601)"
          },
          "fk_series": {
            "allOf": [
              {
                "$ref": "#/components/schemas/idInt"
              }
            ],
            "description": "Foreign key to series.id"
          },
          "name": {
            "type": "string",
            "description": "Season name/number label"
          }
        },
        "required": [
          "id",
          "created_at",
          "fk_series",
          "name"
        ],
        "description": "Seasons table belonging to a series"
      },
      "SeasonListResponse": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/seasonSchema"
        },
        "description": "List of seasons"
      },
      "seasonCreateInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Season name/number label"
          },
          "fk_series": {
            "allOf": [
              {
                "$ref": "#/components/schemas/idInt"
              }
            ],
            "description": "Foreign key to series.id"
          }
        },
        "required": [
          "fk_series",
          "name"
        ],
        "description": "Seasons table belonging to a series"
      },
      "SeriesResponse": {
        "type": "object",
        "properties": {
          "id": {
            "allOf": [
              {
                "$ref": "#/components/schemas/idInt"
              }
            ],
            "description": "Series numeric identifier (PK)"
          },
          "created_at": {
            "allOf": [
              {
                "$ref": "#/components/schemas/timestamp"
              }
            ],
            "description": "Row creation time in UTC (ISO 8601)"
          },
          "name": {
            "type": "string",
            "description": "Series title/name"
          },
          "image_url": {
            "type": "string",
            "description": "Public URL of the series image (nullable)"
          }
        },
        "required": [
          "id",
          "created_at",
          "name",
          "image_url"
        ],
        "description": "Series table with name and optional image URL"
      },
      "SeriesListResponse": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/seriesSchema"
        },
        "description": "List of series"
      },
      "seriesCreateInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Series title/name",
            "nullable": true
          },
          "image_url": {
            "type": "string",
            "description": "Public URL of the series image (nullable)",
            "nullable": true
          }
        },
        "description": "Series table with name and optional image URL"
      },
      "idInt": {
        "type": "integer",
        "minimum": 0,
        "exclusiveMinimum": true
      },
...
      "timestamp": {
        "type": "string"
      },
...
      "seriesSchema": {
        "type": "object",
        "properties": {
          "id": {
            "allOf": [
              {
                "$ref": "#/components/schemas/idInt"
              }
            ],
            "description": "Series numeric identifier (PK)"
          },
          "created_at": {
            "allOf": [
              {
                "$ref": "#/components/schemas/timestamp"
              }
            ],
            "description": "Row creation time in UTC (ISO 8601)"
          },
          "name": {
            "type": "string",
            "description": "Series title/name"
          },
          "image_url": {
            "type": "string",
            "description": "Public URL of the series image (nullable)"
          }
        },
        "required": [
          "id",
          "created_at",
          "name",
          "image_url"
        ],
        "description": "Series table with name and optional image URL"
      },
      "seasonSchema": {
        "type": "object",
        "properties": {
          "id": {
            "allOf": [
              {
                "$ref": "#/components/schemas/idInt"
              }
            ],
            "description": "Season numeric identifier (PK)"
          },
          "created_at": {
            "allOf": [
              {
                "$ref": "#/components/schemas/timestamp"
              }
            ],
            "description": "Row creation time in UTC (ISO 8601)"
          },
          "fk_series": {
            "allOf": [
              {
                "$ref": "#/components/schemas/idInt"
              }
            ],
            "description": "Foreign key to series.id"
          },
          "name": {
            "type": "string",
            "description": "Season name/number label"
          }
        },
        "required": [
          "id",
          "created_at",
          "fk_series",
          "name"
        ],
        "description": "Seasons table belonging to a series"
      },
...
    },
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid request parameters"
                },
                "code": {
                  "type": "string",
                  "example": "400"
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "An unexpected error occurred"
                },
                "code": {
                  "type": "string",
                  "example": "500"
                }
              }
            }
          }
        }
      }
    }
  },
...
  "paths": {
...
    "/seasons": {
      "get": {
        "operationId": "get-seasons",
        "summary": "List all seasons",
        "description": "Fetches all seasons ordered by id",
        "tags": [
          "Seasons"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeasonListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "operationId": "post-seasons",
        "summary": "Create a new season",
        "description": "Creates a season with name and fk_series",
        "tags": [
          "Seasons"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/seasonCreateInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeasonResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/seasons/{id}": {
      "get": {
        "operationId": "get-seasons-{id}",
        "summary": "Get season by id",
        "description": "Fetch a single season by id",
        "tags": [
          "Seasons"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer",
              "description": "Resource ID"
            },
            "required": true,
            "description": "Resource ID",
            "example": 123
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeasonResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "put": {
        "operationId": "put-seasons-{id}",
        "summary": "Update season",
        "description": "Partially update season by id",
        "tags": [
          "Seasons"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer",
              "description": "Resource ID"
            },
            "required": true,
            "description": "Resource ID",
            "example": 123
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeasonResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "delete": {
        "operationId": "delete-seasons-{id}",
        "summary": "Delete season",
        "description": "Delete season by id",
        "tags": [
          "Seasons"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer",
              "description": "Resource ID"
            },
            "required": true,
            "description": "Resource ID",
            "example": 123
          }
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/series": {
      "get": {
        "operationId": "get-series",
        "summary": "List all series",
        "description": "Fetches all series ordered by id",
        "tags": [
          "Series"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "operationId": "post-series",
        "summary": "Create a new series",
        "description": "Creates a series with name and optional image_url",
        "tags": [
          "Series"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/seriesCreateInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/series/{id}": {
      "get": {
        "operationId": "get-series-{id}",
        "summary": "Get series by id",
        "description": "Fetch a single series by id",
        "tags": [
          "Series"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer",
              "description": "Resource ID"
            },
            "required": true,
            "description": "Resource ID",
            "example": 123
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "put": {
        "operationId": "put-series-{id}",
        "summary": "Update series",
        "description": "Partially update series by id",
        "tags": [
          "Series"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer",
              "description": "Resource ID"
            },
            "required": true,
            "description": "Resource ID",
            "example": 123
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "delete": {
        "operationId": "delete-series-{id}",
        "summary": "Delete series",
        "description": "Delete series by id",
        "tags": [
          "Series"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer",
              "description": "Resource ID"
            },
            "required": true,
            "description": "Resource ID",
            "example": 123
          }
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/series/{id}/seasons": {
      "get": {
        "operationId": "get-series-{id}-seasons",
        "summary": "List seasons for a series",
        "description": "Fetches all seasons belonging to a series (hierarchical)",
        "tags": [
          "Series"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer",
              "description": "Resource ID"
            },
            "required": true,
            "description": "Resource ID",
            "example": 123
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeasonListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    }
  }
}